requirements.txt
apache-airflow-providers-apache-spark==4.9.0
apache-airflow==2.9.3
apache-airflow-providers-google==10.1.0
pyspark==3.5.1
pendulum==3.0.0
requests==2.31.0
pandas==2.2.2
scikit-learn==1.5.1
lightgbm==4.5.0
plotly==5.23.0
joblib==1.4.2